refactor: remove dead webview IPC arms
