+++ packages/desktop/src/main/desktopAgentExecution.ts
@@ -834,7 +834,6 @@ export class DesktopProgressBridge {
         },
         file: {
           openFile: (file, line) => this.options.host.openPath(file, line),
-          openFileCompile: (file) => this.openFileCompile(file),
         },
         approval: {
           approvePendingDelegatedWork: (stream, initiatingProposalId) =>
@@ -995,13 +994,6 @@ export class DesktopProgressBridge {
           await this.reportRecordingError(toErrorMessage(error));
           this.postRecordingStatus({ status: 'stopped' });
         }
-      },
-      // Settings navigation
-      openMemoryView: () => {
-        this.showSettings(SETTINGS_TAB.MEMORY);
-      },
-      openProfile: () => {
-        this.showSettings();
       },
       // Pop-out-to-editor is a VS Code editor-tab concept; the desktop app is
       // a single window.
@@ -1251,10 +1243,6 @@ export class DesktopProgressBridge {
         });
       });
     return Promise.resolve();
-  }
-
-  openFileCompile(filePath: string): Promise<void> {
-    return this.fileActions.openFileCompile(filePath);
   }
 
   /**
