+++ packages/desktop/src/main/desktopCrashReporting.ts
@@ -1,70 +1,8 @@
-import escapeRegExp from 'escape-string-regexp';
-import { unique } from '@utils/core';
-import type { ErrorEvent } from '@sentry/electron/main';
-
-const REDACTED_PATH = '<redacted-path>';
+import { createDesktopCrashEventScrubber } from './desktopCrashEventScrubber.js';
 
 export interface DesktopCrashReportingInitOptions {
   sensitivePaths: readonly (string | undefined)[];
   log?: Pick<Console, 'debug' | 'error'>;
-}
-
-function pathVariants(path: string): string[] {
-  const trimmed = path.trim();
-  if (!trimmed) return [];
-  const forward = trimmed.replaceAll('\\', '/');
-  const backward = forward.replaceAll('/', '\\');
-  return unique([trimmed, forward, backward]);
-}
-
-function buildPathScrubbers(paths: readonly (string | undefined)[]): RegExp[] {
-  return paths
-    .flatMap((path) => (path ? pathVariants(path) : []))
-    .filter((path) => path.length > 1)
-    .sort((a, b) => b.length - a.length)
-    .map((path) => new RegExp(escapeRegExp(path), 'gi'));
-}
-
-function scrubString(value: string, scrubbers: readonly RegExp[]): string {
-  return scrubbers.reduce(
-    (current, scrubber) => current.replace(scrubber, REDACTED_PATH),
-    value,
-  );
-}
-
-function scrubValue(value: unknown, scrubbers: readonly RegExp[]): unknown {
-  if (typeof value === 'string') {
-    return scrubString(value, scrubbers);
-  }
-  if (Array.isArray(value)) {
-    return value.map((item) => scrubValue(item, scrubbers));
-  }
-  if (value && typeof value === 'object') {
-    return Object.fromEntries(
-      Object.entries(value).map(([key, entry]) => [
-        scrubString(key, scrubbers),
-        scrubValue(entry, scrubbers),
-      ]),
-    );
-  }
-  return value;
-}
-
-export function scrubDesktopCrashEvent(
-  event: ErrorEvent,
-  sensitivePaths: readonly (string | undefined)[],
-): ErrorEvent | null {
-  return createDesktopCrashEventScrubber(sensitivePaths)(event);
-}
-
-function createDesktopCrashEventScrubber(
-  sensitivePaths: readonly (string | undefined)[],
-): (event: ErrorEvent) => ErrorEvent | null {
-  const scrubbers = buildPathScrubbers(sensitivePaths);
-  return (event) => {
-    if (event.platform !== 'native') return null;
-    return scrubValue(event, scrubbers) as ErrorEvent;
-  };
 }
 
 /**
