+++ src/test-kernel/desktop/DesktopCrashReporting.vitest.ts
@@ -1,50 +1,38 @@
-import { beforeAll, describe, expect, it } from 'vitest';
-
-import { loadSourceModule } from './loadSourceModule.ts';
-
-type ScrubDesktopCrashEvent =
-  (typeof import('@desktop/main/desktopCrashReporting'))['scrubDesktopCrashEvent'];
+import { describe, expect, it } from 'vitest';
+import { createDesktopCrashEventScrubber } from '@desktop/main/desktopCrashEventScrubber';
 
 describe('desktop crash reporting', () => {
-  let scrubDesktopCrashEvent: ScrubDesktopCrashEvent;
-
-  beforeAll(async () => {
-    ({ scrubDesktopCrashEvent } = await loadSourceModule(
-      '@desktop/main/desktopCrashReporting',
-    ));
-  });
-
   it('drops non-native events for the v1 crash-reporting scope', () => {
-    expect(
-      scrubDesktopCrashEvent({ type: undefined, platform: 'javascript' }, []),
-    ).toBeNull();
+    const scrub = createDesktopCrashEventScrubber([]);
+    expect(scrub({ type: undefined, platform: 'javascript' })).toBeNull();
   });
 
   it('scrubs workspace and app paths from native crash events', () => {
-    const scrubbed = scrubDesktopCrashEvent(
-      {
-        type: undefined,
-        platform: 'native',
-        message: 'Crash in /Users/alice/paper/main.tex',
-        exception: {
-          values: [
-            {
-              stacktrace: {
-                frames: [
-                  { filename: 'C:\\Users\\Alice\\TeXRA\\workspace\\agent.ts' },
-                ],
-              },
+    const scrub = createDesktopCrashEventScrubber([
+      '/Users/alice/paper',
+      'C:\\Users\\Alice\\TeXRA\\workspace',
+    ]);
+    const scrubbed = scrub({
+      type: undefined,
+      platform: 'native',
+      message: 'Crash in /Users/alice/paper/main.tex',
+      exception: {
+        values: [
+          {
+            stacktrace: {
+              frames: [
+                { filename: 'C:\\Users\\Alice\\TeXRA\\workspace\\agent.ts' },
+              ],
             },
-          ],
-        },
-        extra: {
-          cwd: '/Users/alice/paper',
-          log: 'C:\\Users\\Alice\\TeXRA\\workspace\\logs\\main.log',
-          '/Users/alice/paper/output.pdf': 'build output',
-        },
+          },
+        ],
       },
-      ['/Users/alice/paper', 'C:\\Users\\Alice\\TeXRA\\workspace'],
-    );
+      extra: {
+        cwd: '/Users/alice/paper',
+        log: 'C:\\Users\\Alice\\TeXRA\\workspace\\logs\\main.log',
+        '/Users/alice/paper/output.pdf': 'build output',
+      },
+    });
 
     const serialized = JSON.stringify(scrubbed);
     expect(serialized).toContain('<redacted-path>');
